Refactor execution
